AtomicAccees now works for 64 bit. Bugfix synchronized_ObjectJc itself was not thread safe. * OsWrapperJc: Now some things with atomic access, before: synchronized_ObjectJc was not thread safe! * Fwc/objectBaseC: The idSyncHandles is now defined in a 8-Byte struct State_ObjectJc.   From this struct the 32-bit word (aligned 4) should be set with atomic access to store idSyncHandles. * Fwc/objectBaseC: isInitialized_ObjectJc() and setInitialized_ObjectJc, instead immediately using the bit.   Refactored in all using sources. * Inspc: Some advancement for usage in Simulink.

+++ source/Fwc/fw_handle_ptr64.h
@@ -66,7 +66,7 @@ const char* init_Handle2Ptr(int nrofEntries);
   /**Invocation of INIT is empty because there is not Handle2Ptr. */
   #define INIT_Handle2Ptr(nrofEntries)  
   #define setPtr_Handle2Ptr(PTR, NAME, H) (*H=PTR, null)
-  #define handle_Handle2Ptr(PTR, H) (*H=PTR, null)
+  #define handle_Handle2Ptr(PTR, H) (*H=(uint32)PTR, null)
 #endif
